get_layman_users does not return LAYMAN_GS_USER anymore
